requirements.txt
attrs==21.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.0.4
distlib==0.3.6
filelock==3.12.0
identify==2.5.23
idna==3.2
iniconfig==1.1.1
isort==5.12.0
nodeenv==1.7.0
packaging==21.0
platformdirs==3.4.0
pluggy==0.13.1
pre-commit==3.2.2
py==1.11.0
pyparsing==2.4.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.22.0
